+++ src/Benchmarks/Startup.cs
@@ -2,8 +2,12 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
+using System.Collections;
 using System.Data.Common;
 using System.Data.SqlClient;
+using System.Diagnostics;
+using System.IO;
+using System.Reflection;
 using System.Text.Encodings.Web;
 using System.Text.Unicode;
 using Benchmarks.Configuration;
@@ -298,6 +302,42 @@ namespace Benchmarks
             }
 
             app.RunDebugInfoPage();
+
+
+            var runtimeInfo = Configuration["runtimeInfo"];
+            if (runtimeInfo != null && bool.Parse(runtimeInfo))
+            {
+                DisplayRuntimeInfo();
+            }
+        }
+
+
+        private static void DisplayRuntimeInfo()
+        {
+            Console.WriteLine("");
+            Console.WriteLine("Environment variables");
+            Console.WriteLine("-----------------------");
+
+            foreach (DictionaryEntry entry in Environment.GetEnvironmentVariables())
+            {
+                Console.WriteLine($"{entry.Key}={entry.Value}");
+            }
+
+            Console.WriteLine("");
+            Console.WriteLine("Loaded modules");
+            Console.WriteLine("-----------------------");
+
+            foreach (var m in Process.GetCurrentProcess().Modules.OfType<ProcessModule>())
+            {
+                Assembly assembly = null;
+                try
+                {
+                    assembly = Assembly.Load(new AssemblyName(Path.GetFileNameWithoutExtension(m.ModuleName)));
+                }
+                catch { }
+
+                Console.WriteLine($"{m.FileName} {m.ModuleName} {assembly?.GetName().Version.ToString()} {assembly?.GetCustomAttribute<AssemblyInformationalVersionAttribute>()?.InformationalVersion}");
+            }
         }
     }
 }
